requirements.txt
# ── Core framework ────────────────────────────────────────────────────────────
streamlit>=1.32.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.11

# ── Google Gemini ─────────────────────────────────────────────────────────────
google-generativeai>=0.4.0

# ── Document processing ───────────────────────────────────────────────────────
pdfplumber>=0.10.0
PyPDF2>=3.0.0
python-docx>=1.0.0
python-pptx>=0.6.23

# ── Vector DB & embeddings ────────────────────────────────────────────────────
faiss-cpu>=1.7.4
langchain-community>=0.0.20

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
